requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.5.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pymupdf>=1.22.0
pdfminer.six>=20221105
requests>=2.28.0